Move cesr_adapter out of event_parsing module
